with db

+++ app/src/main/java/com/example/contacts/Contact.kt
@@ -1,4 +1,4 @@
 package com.example.contacts
 
 
-data class Contact(val contactID:Int,val contactName:String?,val contactPhoneNumber:List<String>?,val contactEmail:List<String>?,val contactAddress:List<String>?)
+data class Contact(val contactID:Long,val contactName:String?,val contactPhoneNumber:List<String>?,val contactEmail:List<String>?,val contactAddress:List<String>?)
